requirements.txt
fastapi
uvicorn
jinja2
newspaper3k
beautifulsoup4
requests
reportlab
lxml_html_clean
